Revert "wpcom-proxy-request: Return Promise if no callback specified (Take Two) (#39722)"
This reverts commit 8ff15c65b230808c0107a4c687f8c57438cd0153.
