--- packages/data-stores/src/shared-types.ts
@@ -1,4 +1,0 @@
-export interface WpcomClientCredentials {
-	client_id: string;
-	client_secret: string;
-}
