+++ packages/wpcom-proxy-request/History.md
@@ -1,9 +1,5 @@
 # 6.0.0 / TBD
 
-- Breaking: Return Promise (rather than `XMLHttpRequest` instance) if no callback argument is provided.
-  - In practice, most people have probably been using the callback rather than the returned `XMLHttpRequest` instance, so this shouldn't be a breaking change for most.
-- Add `requestAllBlogsAccess()`.
-- Add a few type definitions.
 - Move the published `build/` folder to `dist/` to align with other Calypso packages
 - Upgrade dependency 'debug' to 4.1.1
 - Remove `component-event` dependency, use `addEventListener`/`removeEventListener` directly
