+++ packages/wpcom-proxy-request/package.json
@@ -30,11 +30,9 @@
 	},
 	"files": [
 		"dist",
-		"types",
 		"History.md",
 		"README.md"
 	],
-	"types": "types",
 	"scripts": {
 		"clean": "npx rimraf dist",
 		"prepublish": "npm run clean",
